apps/microservices/heimdall/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "heimdall" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/apps/microservices/heimdall/html' --cov-report xml:'../../../coverage/apps/microservices/heimdall/coverage.xml' --html='../../../reports/apps/microservices/heimdall/unittests/html/index.html' --junitxml='../../../reports/apps/microservices/heimdall/unittests/junit.xml'"

[tool.poetry]
name = "heimdall"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

[tool.isort]
known_third_party = ["heimdall."]

[[tool.poetry.packages]]
include = "heimdall"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = "^4.2.3"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
pyjwt = "^2.7.0"
drf-spectacular = "^0.26.3"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
whitenoise = "^6.5.0"
google-auth = "^2.22.0"
opentelemetry-api = "1.27.0"
opentelemetry-sdk = "1.27.0"
opentelemetry-exporter-otlp-proto-http = "1.27.0"
opentelemetry-instrumentation-django = {version = "0.48b0", extras = ["asgi"]}
psycopg2 = "^2.9.9"
django-allauth = {extras = ["socialaccount"], version = "^64.2.1"}
atlas-provider-django = "^0.2.1"
opentelemetry-instrumentation-asgi = "^0.48b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
